requirements.txt
boto3==1.13.3
botocore==1.16.3
docutils==0.15.2
jmespath==0.9.5
python-dateutil==2.8.1
six==1.14.0
urllib3==1.25.9
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
requests-aws4auth==1.0.1
python-geohash==0.8.5
pytz==2020.4
